Update Node.js version in GitHub Actions workflow
Updated the Node.js version used in the GitHub Actions workflow from 16.x to 20.x. This change is reflected in the `markdownlint.yml` file under the `Use Node.js` step. The specific line of code that was changed is `node-version: 16.x` which has been updated to `node-version: 20.x`. This update could potentially affect the compatibility and performance of the workflow, depending on the Node.js features being used.