Updated user profile editing and service
Updated the URL for the user profile edit page and made changes to the `EditProfile.razor` page to allow email editing. The `UpdateProfile` method in `EditProfile.razor` is now public and updates the user profile using the `UserProfileService`. The `UserProfileService` has been refactored to remove the dependency on `HttpClient` and now uses the `AuthenticationStateProvider`. Two new methods, `GetProfileData` and `UpdateProfileData`, have been added to `UserProfileService` to handle user profile data. The `UserProfile` class has been updated to remove the `EmailVerified` property and add a `UserId` property.