Replace image logos with SVG text in header and footer
Replaced the image logos in the header and footer of the website with SVG text logos. The changes were made in `FooterBar.razor` and `HeaderBar.razor`. An additional image tag without a source attribute was added to `FooterBar.razor`, which might be an error as it will not display anything.