Refactor `GetProfileDataAsync` in `ProfileService` class
Significant refactoring has been done to the `GetProfileDataAsync` method in the `ProfileService` class. The retrieval of the `user` object has been changed to include null checking and a different method for finding the user. The creation of the `claims` list now uses the `GetClaimsFromUser` method. Role claims are now added using the `Select` LINQ method and `AddRange`. The previous code for creating and populating the claims list and adding role claims has been replaced with this new approach.