Subject: Update `name` variable assignment in `CascadeAuthState.razor`
Body:
Updated the assignment of the `name` variable in `CascadeAuthState.razor` to prevent potential `NullReferenceException`. The new code uses the null-conditional operator to check if the user has a claim of type "name". If not, it assigns an empty string to `name` using the null-coalescing operator, ensuring `name` is always assigned a string value.